target move action added

+++ src/Grpc/GrpcMobile.cs
@@ -6,7 +6,7 @@ namespace ClassicUO.Grpc
     internal class GrpcMobile
     {
         public string Name;
-        public uint X;
-        public uint Y;
+        public float X;
+        public float Y;
     }
 }
